Rename FileDetails -> ValidationRequest
