+++ src/main/java/uk/gov/companieshouse/account/validator/controller/AccountValidationController.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import uk.gov.companieshouse.account.validator.message.ResponseMessage;
-import uk.gov.companieshouse.account.validator.model.FileDetails;
+import uk.gov.companieshouse.account.validator.model.validation.ValidationRequest;
 import uk.gov.companieshouse.account.validator.service.AccountValidatedService;
 import uk.gov.companieshouse.account.validator.service.impl.AccountValidatorImpl;
 import uk.gov.companieshouse.logging.Logger;
@@ -31,10 +31,10 @@ public class AccountValidationController {
     }
 
     @PostMapping("/validate")
-    public ResponseEntity<ResponseMessage> validate(@Valid @RequestBody FileDetails fileDetails) {
+    public ResponseEntity<ResponseMessage> validate(@Valid @RequestBody ValidationRequest validationRequest) {
         String message = "";
         try {
-            boolean result = accountValidatorImpl.downloadIxbrlFromLocation(fileDetails);
+            boolean result = accountValidatorImpl.downloadIxbrlFromLocation(validationRequest);
             if (result) {
                 message = "File validated successfully";
                 return ResponseEntity.status(HttpStatus.OK).body(new ResponseMessage(message));
@@ -43,7 +43,7 @@ public class AccountValidationController {
                 return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED).body(new ResponseMessage(message));
             }
         } catch (Exception e) {
-            message = "Could not validate the file: " + fileDetails.getFile_name() + ". Error: " + e.getMessage();
+            message = "Could not validate the file: " + validationRequest.getFile_name() + ". Error: " + e.getMessage();
             return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED).body(new ResponseMessage(message));
         }
     }
